Fix score_percent ClassCastException (Double vs Integer)
PHP was returning score_percent as a PHP float (round($x, 2)), which
json_encode could encode as a JSON float. Gson occasionally deserialised
that into a boxed Double, causing getIntExtra("score_percent") to
throw a ClassCastException and show 0% on the results screen.

Fix: make the type consistently integer end-to-end:
- submit_quiz.php: (int) round($scorePercent) → emits JSON integer
- QuizSubmitResponse: double scorePercent → int scorePercent
- QuizActivity: remove now-redundant (int) cast on getScorePercent()

+++ app/src/main/java/com/example/literise/activities/QuizActivity.java
@@ -293,7 +293,7 @@ public class QuizActivity extends AppCompatActivity {
         intent.putExtra("node_id", nodeId);
         intent.putExtra("module_id", moduleId);
         intent.putExtra("lesson_number", lessonNumber);
-        intent.putExtra("score_percent", (int) result.getResult().getScorePercent());
+        intent.putExtra("score_percent", result.getResult().getScorePercent());
         intent.putExtra("correct_count", result.getResult().getCorrectCount());
         intent.putExtra("total_questions", result.getResult().getTotalQuestions());
         intent.putExtra("adaptive_decision", result.getResult().getAdaptiveDecision());
